# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kamwaalay API",
        "description": "API documentation for Kamwaalay - A platform connecting households with trusted domestic helpers",
        "contact": {
            "email": "contact@kamwaalay.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "API Server"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/onboarding/helper": {
            "get": {
                "tags": [
                    "Onboarding"
                ],
                "summary": "Get helper onboarding form",
                "description": "Get form data for helper onboarding. Requires helper role.",
                "operationId": "9a99982590a7317e16ad6506da95d863",
                "responses": {
                    "200": {
                        "description": "Onboarding form data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Only helpers can access"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Onboarding"
                ],
                "summary": "Complete helper onboarding",
                "description": "Complete helper onboarding by submitting services, NIC document, and profile information. Requires helper role.",
                "operationId": "8cd2b30316330d4cfa8f937675555116",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "services",
                                    "nic",
                                    "nic_number"
                                ],
                                "properties": {
                                    "services": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "service_type",
                                                "work_type",
                                                "location_id"
                                            ],
                                            "properties": {
                                                "service_type": {
                                                    "type": "string",
                                                    "enum": [
                                                        "maid",
                                                        "cook",
                                                        "babysitter",
                                                        "caregiver",
                                                        "cleaner",
                                                        "all_rounder"
                                                    ]
                                                },
                                                "work_type": {
                                                    "type": "string",
                                                    "enum": [
                                                        "full_time",
                                                        "part_time"
                                                    ]
                                                },
                                                "location_id": {
                                                    "type": "integer"
                                                },
                                                "monthly_rate": {
                                                    "type": "number",
                                                    "minimum": 0,
                                                    "nullable": true
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "maxLength": 2000,
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "nic": {
                                        "description": "NIC document file (jpeg, jpg, png, pdf, max 5MB)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "nic_number": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "photo": {
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    },
                                    "skills": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "experience_years": {
                                        "type": "integer",
                                        "minimum": 0,
                                        "nullable": true
                                    },
                                    "bio": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Onboarding completed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Onboarding completed successfully!"
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Only helpers can complete onboarding"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/onboarding/business": {
            "get": {
                "tags": [
                    "Onboarding"
                ],
                "summary": "Get business onboarding form",
                "description": "Get form data for business onboarding. Requires business role.",
                "operationId": "de8707ff5d1b87174a3991b30ce27736",
                "responses": {
                    "200": {
                        "description": "Onboarding form data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Only businesses can access"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Onboarding"
                ],
                "summary": "Complete business onboarding",
                "description": "Complete business onboarding by submitting business information. Requires business role.",
                "operationId": "d9121b7d2b0a0f274c604f0c0c40ea18",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "business_name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "nullable": true
                                    },
                                    "business_address": {
                                        "type": "string",
                                        "maxLength": 500,
                                        "nullable": true
                                    },
                                    "business_description": {
                                        "type": "string",
                                        "maxLength": 2000,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Onboarding completed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Onboarding completed successfully!"
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Only businesses can complete onboarding"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            },
            "0": []
        }
    },
...
}